pyproject.toml
[tool.poetry]
name = "poster"
version = "0.1.0"
description = ""
authors = ["Wasil P <it.dzemidok@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
